Refactors pdf generation
