+++ apps/judicial-system/backend/src/app/formatters/writeFile.ts
@@ -1,0 +1,6 @@
+import fs from 'fs'
+
+export function writeFile(fileName: string, documentContent: string) {
+  // In e2e tests, fs is null and we have not been able to mock fs
+  fs?.writeFileSync(`../${fileName}`, documentContent, { encoding: 'binary' })
+}
